Add symetrical drawing
View->Symmetry, bound to the pipe character "|".  Drawing and straight lines
are mirrored over the x axis in realtime.

TiledSurface has new variables do_symmetry and center_x.

center_x is set to the center of the drawing area.  This is dependent on the
size of the window, so resizng will change the center axis.

Closes bug: https://gna.org/bugs/?func=detailitem&item_id=18375