brushlib: fix smudge starting transparent
Fix a problem where a smudging stroke would suddenly start
with transparency, even in the center of a solid layer.

https://gna.org/bugs/index.php?20034

Fixed by reordering of code: update the smudge color before using
it, instead of after. I'm not sure why it wasn't ordered like that
before, I see no advantage to the old order.

After calling reset() on a brush, the smudge color is set to
zero (transparent). Use the LAST_GETCOLOR_RECENTNESS state (which
is also reset to zero) to detect the initial get_color().

Reordering changes the behaviour of all smudge and watercolor brushes
very slightly, in theory. In practice I was unable to notice any
difference at all.

MyPaint still has a tendency to round down the alpha channel and
thus introduce slight transparency when smudging the same spot really
often, but it is a different problem.