brushlib: Fix resizing of operation queue causing line gaps
The previous code for copying the values in a tile map to a bigger
map during resize did not take into account that a tile
would be stored in a different position in the bigger map.
Thus when processing after a op queue resize, the dirty tile list
would point to wrong, empty queues. This caused gaps in drawn lines
the first time they exceeded the op queue size.

Sorry for big commit. Also makes all functions that have
to do with the tile map operate on a TileMap instead of OperationQueue.