brushlib: new tilt calculation
Tilt value did vary (incorrectly) with ascension.

https://gna.org/bugs/?21735

Quote from http://sourceforge.net/p/linuxwacom/mailman/linuxwacom-devel/?viewmonth=201203

"GIMP, Inkscape, and MyPaint all treat the values they get from GTK as
the normalized length of the tool's projection along the corresponding
axis (they take the arctangent of the values to get an azimuth). While
GTK doesn't explicitly state that this is what applications receive,
its a reasonable reading of their documentation: "tilt attributes
range from -1.0 to 1.0 ... representing the maximum tilt to the left
or up ... [and] to the right or down". It looks like in practice GTK
actually only scales the tilt values it gets from X based on the
minimum and maximum. Its not what applications expect, but appears to
work well enough that nobody's bothered to file a bug."