i18n: update po/*.po
Run scons translate=all. No time like the present.

libmypaint is stable enough to do this in advance of the freeze for MyPaint
itself, and if we are to move to a continuous translation workflow per
mypaint/mypaint#348 we'll need to do this fairly regularly anyway.

In particular "Anti-aliasing" got renamed to "Pixel feather", and some of
the translations may not have this yet.

Translators should aso watch out for formatting issues for those settings
with lists of numbers and their meanings. Looks like this isn't an
intltool-update artefact.