Port libmypaint to autotools.
Porting all current functionnalities of the scons build to autotools.
Some things are still shady and actually some options (or combinations
of options) are broken with the scons build, so it is sometimes hard
to compare if this does 100% the same thing. But that's a good start
to go from.