AC_OPENMP is available as a base autoconf macro since version 2.62.
AX_OPENMP was not a typo (cf. commit 0085d06). It was an external
autoconf macro. My error was that it was already installed on my system
through the Autoconf Macro Archive.
Thus I failed to see it would not work for someone who don't have it
installed. I should have copied it manually into the build system.
Anyway AC_OPENMP is apparently more efficient on some platforms and
autoconf 2.62 is a years-old release (2008), so let's go and assume that
everybody who wants to build libmypaint have a newer version.