Add -lm to relevant _LIBADD
libmypaint shouldn't require programs using it to link against libm;
that should be explicitly resolved at runtime.

Hope this is the right way to do it. It certainly fixes a warning in my
initial test Debianization, as mentioned in
https://lists.debian.org/debian-devel-announce/2011/02/msg00011.html :

> A common mistake with automake is to add libraries to the various LDFLAGS
> macros.  If you need to link a program with a library, add this library
> to one of the LDADD macros, if you need to link a library with a library,
> add to one of the LIBADD macros.  For tracking dependencies with local
> libraries, you may need add the library to one of the DEPENDENCIES
> macros.  See the automake manual for more details.

Some confirmation: libm.so is now listed in the output of
objdump -p libmypaint.so | grep NEEDED