Separate mypaint-config.h and config.h.
MyPaint uses some of the constants from `mypaint-config.h` but much of
the generated part from configuration should not be public because it
conflicts with the configuration of other autotooled project.
Split it so that the generated constants are not in the public header.
This means also that config.h cannot be included from any of the other
public headers and the main problem with this was that the macro
MYPAINT_CONFIG_USE_GLIB cannot be used inside a public header anymore.
I slightly reorganized installed files to account for this.