libtool: API platform and ABI version decls
Update the up-front version macro names to distinguish ABI and API
versions.

We have a concept of a platform version now, for use in .so, .typelib,
.gir names, and in the introspection namespace.

Use --version-info appropriately.

Use --release so that multiple versions of libmypaint can, in theory, be
installed together. Not going the route of declaring separate
libmypaint_1_3_la_LDFLAGS etc just yet
(cf. https://autotools.io/libtool/version.html)
There are indeed drawbacks with not doing this, but we don't build any .a
file, and Debian policy is not to install .la files without a good
runtime linker reason. Also I hate not being able to AC_SUBST the things.