Ascension values can go out of bounds/ add console print inputs
MYPAINT_BRUSH_STATE_ASCENSION seems to grow without bounds as it is constantly having values added and subtracted.  If this number becomes negative than the formula fails.  Constraining MYPAINT_BRUSH_STATE_ASCENSION] with abs() seems to fix the issue, but I have to wonder if it should be allowed to grow and grow over the course of a drawing.  You can basically wind it up or down by spinning your stylus like the hands on a clock.  I've added the remaining inputs to the console for outputting so it is easier to debug.  I don't see why we wouldn't want all the inputs printed to console when you've selected