Input Calibration: viewzoom and viewrotation inputs
When rotating canvas or zooming in, inputs such as speed, direction,
etc should continue to work as if you just rotated a physical canvas
in front of you or peered closer at your work.  As it stands, rotating
the view or zooming in/out will change the inputs in an (arguably)
unexpected way.

Viewzoom is made a useable input for end-users, as it can be useful
for certain inputs such as changing brush radius as you zoom in.

Closes mypaint/libmypaint#66