Smudge: RGB->RYB mixing
This patch adds an RGB to RYB step before and after the normal
smudge mixing code.  The RYB helper was created from information
on the site here:http://www.deathbysoftware.com/colors/index.html

This patch allows intuitive mixing such as yellow+blue=green as well
as desaturation when mixing subtractive complementary colors like
red+green. A further adjustment is made to desaturate disimilar
hues progressively as they are father apart on the wheel.  This
is an attempt to model the nature of pigments: any mix of paints
will reduce chroma.