GridMap: New input/settings for textures and canvas effects
Somewhat like Stroke and Random, provides an oscillating input from
0-256.  These values are based on the X/Y location, so they are
persistent regardless of time or direction of stroke.  The X and Y
location has modulo 256 performed on it, producing the input value.

Also included are settings to affect the scale of this gridmap.
The base scale can be set on a log scale and then modified in either
the X or Y axis by a scale of 0X-10X.

(cherry picked from commit dbc0f5a0f862729ba234b4760667b114bec4bdef)