Boolean #define-s removed from json-c in 2017.
We don't need to #undef TRUE and FALSE before including json.h as this
redefinition has been removed in upstream json-c.
See json-c commit 0992aac61f8b087efd7094e9ac2b84fa9c040fcd.

Note that it still works even if using an older json-c since the
original code was alreadying #undef-ing these 2 values before redefining
them. So that was anyway useless code and this change should work both
with old or new json-c.

Note: the previous code was still working fine when building on Linux,
but not when cross-compiling on Windows. But this is the correct fix
anyway.