Fix never-ending smudge
The consequence of never letting the alpha go below
a certain value when updating the recorded sampled color
(when updating the smudge state) is that the smudge never
stops once it has recorded any non-transparent color.

Another consequence is that never dropping a color when
smudging across transparent areas causes the color to
re-emerge when crossing into opaque areas - probably not
expected behavior.

Better to try to solve the color noise problem in the
blending function.