Use legacy smudging when paint is constant 0
When the paint setting is completely disabled
(set to 0 with no dynamics) just use the old
smudging - including color sampling.

This may create some problems for a couple of dabs
when switching from a brush using spectral smudging
and a brush using legacy smudging, but if that turns
out to be a problem, we can check for state changes
and clear existing smudge data accordingly.