Expand spectral smudge exceptions to dark colors
The last remaining strong artifacts are caused by colors
that are very nearly black. This handles that by falling
back to additive blending if either input or output are
very dark.

Note that the second comparison compares a premultiplied
color, which is not quite correct, but still appears to
work fine.