Fix smudge bugs when using lock alpha
Don't run the x_LockAlpha_y functions when the output
alpha is 0 (pure erasing). Finding this bug was made a lot
easier due to the setting of colors to bright red/green
when those colors should not be visible in the output.

Add back the low-alpha noise protection without retaining
outdated color information; since we are storing straight
color channels we can just update the alpha channel of
the existing smudge color instead of running the spectral
mixing. The paint factor becomes irrelevant in this case,
as no colors are mixed.