Make spectral color sampling scale linearly
To counteract the rounding errors that build up from
repeatedly using fastpow in partial results, let the
sample percentages be inversely proportional to the
brush radius.

Another way to mitigate the rounding errors (to some extent)
is to use powf instead of fastpow, but reducing the samples
is the more performant approach.

Note that this does _not_ completely remove accumulated errors,
it only makes it a lot harder to produce them by accident.