Enable use of multiple invalidation rectangles
THIS UPDATE CHANGES THE API!
See changes to mypaint_surface_end_atomic signature/implementations.

Calling code can still use a single rectangle for updates,
but instead of a plain MyPaintRectangle pointer, it must use
a MyPaintRectangles struct where the field num_rectangles is
set to 1, and the rectangles field points to enough memory
to contain a single MyPaintRectangle struct.

In order to support efficient redraws when dabs are spread out across
wide areas, but with relatively small areas that are actually changed,
we now track multiple rectangles (referred to as bounding boxes) that
are then sent back to the calling code on completion of a drawing op.

For now, only symmetry modes make use of multiple bounding boxes,
though it might be feasible to extend the use to dabs that are placed
through the offset settings.

This commit refactors the symmetry code, but leaves the math as it was.