Fix capitalization for unmerged strings [skip ci]
The names of the settings were all capitalized, which is arguably
more correct than the current style, but still creates a style
clash that can look ugly and (more importantly) can be confusing
for translators.

Capitalization is removed for new settings that will be placed
in the same category as existing settings with multi-word names.

Capitalization is retained for new settings with their own categories.
It is also retained for all new inputs, since there were previously
no multi-word input names that would cause a clash.

In the long term, we'll probably want to make capitalized
settings names the standard, but now is not the time to add
to the workload of the translators.