Refactor symmetry code and add angle parameter
THIS COMMIT CHANGES THE API:
the 'set_symmetry_state' function in mypaint-tiled-surface.h
now takes an additional angle parameter as its 4th argument.
The function is not part of the official API, but nevertheless
exposed.

---

Create new structs to hold information and data necessary
to perform symmetry calculations using matrix transforms.
Create new functions to initialize and update that data.

Make it possible to adjust the angle for the symmetry modes,
tilting the reference frame counterclockwise (unit = degrees).

Transformation matrices are precalculated and stored when required
(right before they are used, if the symmetry parameters have changed).

NOTE: this code tries to allocate more matrices if required, without
control from the caller. This is generally bad practice, and should
be fixed at some point (the same is already done for bounding boxes,
and in other older code as well).

Additionally: fixes incorrect dab angle being used for the dabs
mirrored across the horizontal line.