Add softness setting
Committer's comment:

The softness setting provides a way to linearly scale the opacity
values of the dab masks, as calculated base on the hardness setting.

As such, it is functionally very similar to "Opacity multiply", but
without the drawback of having to deal with positive values resulting
in always-on strokes, even at 0 pressure (which can be avoided to some
extent, but is a hassle for brush makers).