Move smudge buckets into MyPaintBrush struct
MyPaint currently only uses a single instance of MyPaintBrush, but
GIMP sometimes uses multiple, so the smudge state cannot be global.

A separate constructor allocates a given number of  smudge buckets.
For the default constructor, the corresponding segment of the state
array is used instead of separately allocated buckets.

The range of buckets that are actually used between resets are kept
track of, to avoid unnecessarily large memsets. Could be made more
sophisticated, but we'll assume that brushes making use of smudge
buckets will avoid large gaps in use of buckets.