Use old dab count interpolation in legacy strokes
The mypaint_brush_stroke_to function now uses the old way of
calculating the number of dabs to use for strokes, meaning that
there is no need for an initial call to account for an empty segment.

The mypaint_brush_stroke2 function still uses the state values instead
of base values to calculate the dabs, but falls back to using base values
for the intial call.