i18n: Add/change scripts to fix source string locs
Previously the source string locations pointed to the unreadable
lines in a non-tracked generated header, making the source string
location useless. This changes the script generating the header
to include information about the real source location for the strings,
which is subsequently turned into an accurate source location in the
script generating the po template file.