brushmodes: move constant calculations out of loop
Moving out the rgb to spectral conversion for the input color
yields overall speedups of 3-5% (depending on the complexity
of the stroke interpolation - #mappings etc). The compiler is
not able to infer the constant'ness of the output from the
conversion call.

All credit to SleepProgger for noticing that this was not done in
the regular and alpha-lock versions of the spectral blend functions.