Add new stroke_to function & bump version to 1.6.0
When the colors drawn on the surface are interpreted as linear
sRGB, the regular HSV/HSL color dynamics produce values that
don't match that space.

The new public function mypaint_stroke_to_2_linearsRGB handles
that by transforming the colors channels before and after the
HSV/HSL adjustments (if at least one of them is used).