Stop using `-release`, reset abi version fields
The use of `-release` forces recompilation of reverse dependencies
on every new release and makes the abi version fields essentially
cosmetic (even if they correctly track (in)compatibility).

Clarified the comments on how to handle the abi fields on releases,
since the short forms didn't cover resetting the age/revision flags.

Reordered the field declarations in the automake file to match the
order in which they are passed to -version-info.